Move DB ops to separate directory
